Option to retain initial time (like deSolve)
This should be dropped when speed is the main thing needed (in many
applications the initial condition is never referred to again as the
final condition is what is wanted) but when transient dynamics are wanted
this is useful.  Note that this involves calling the output function
an additional time (if enabled).