Big rework of argument defaults
Move towards deSolve compatibility in outputs, by default.  This makes
the minimal output more opt-in.  This should be less disruptive and work
better with projects that might want to use both.  Hopefully I don't
regret doing this because it will be very difficult to wind this back
(look at the number of changed tests)