Beginnings of event support
This is *far* from complete, but sketches out some of the key ideas.
We're going to drive events off of critical times because we're
already planning on stopping there and that simplifies the bookkeeping
within the main integration function significantly.

events at the moment can only modify the state vector and not modify
the actual data, but we'll get that relaxed soon.  Lots of other
restrictions about the style of events, mixing C and R, etc that will
be resolved in the next few commits.